+++ src/types/eslint-js.d.ts
@@ -1,5 +1,6 @@
 declare module '@eslint/js' {
   import type { TSESLint } from '@typescript-eslint/utils';
+
   const value: TSESLint.FlatConfig.Plugin & {
     configs: {
       recommended: TSESLint.FlatConfig.Config;