+++ src/types/eslint-plugin-jest-dom.d.ts
@@ -1,5 +1,6 @@
 declare module 'eslint-plugin-jest-dom' {
   import type { TSESLint } from '@typescript-eslint/utils';
+
   const value: TSESLint.FlatConfig.Plugin & {
     configs: {
       'flat/all': TSESLint.FlatConfig.Config;