+++ src/types/eslint-plugin-mocha.d.ts
@@ -1,5 +1,6 @@
 declare module 'eslint-plugin-mocha' {
   import type { TSESLint } from '@typescript-eslint/utils';
+
   const value: TSESLint.FlatConfig.Plugin & {
     configs: {
       recommended: TSESLint.FlatConfig.Config;