+++ src/types/eslint-plugin-next.d.ts
@@ -1,5 +1,6 @@
 declare module '@next/eslint-plugin-next' {
   import type { TSESLint } from '@typescript-eslint/utils';
+
   const value: TSESLint.FlatConfig.Plugin & {
     configs: {
       recommended: TSESLint.FlatConfig.Config;