+++ src/types/eslint-plugin-react.d.ts
@@ -1,5 +1,6 @@
 declare module 'eslint-plugin-react' {
   import type { TSESLint } from '@typescript-eslint/utils';
+
   const value: TSESLint.FlatConfig.Plugin & {
     configs: {
       all: TSESLint.FlatConfig.Config;