+++ src/types/eslint-plugin-testing-library.d.ts
@@ -1,5 +1,6 @@
 declare module 'eslint-plugin-testing-library' {
   import type { TSESLint } from '@typescript-eslint/utils';
+
   const value: TSESLint.FlatConfig.Plugin & {
     configs: {
       'flat/angular': TSESLint.FlatConfig.Config;