+++ src/types/eslint-plugin-unicorn.d.ts
@@ -1,5 +1,6 @@
 declare module 'eslint-plugin-unicorn' {
   import type { TSESLint } from '@typescript-eslint/utils';
+
   const value: TSESLint.FlatConfig.Plugin;
   export default value;
 }